user can delete only their own blog

+++ osa4/controllers/blogs.js
@@ -56,6 +56,24 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/delete/:id', async (request, response, next) => {
   try {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(400).json({ error: 'token missing or invalid' })
+    }
+
+    const userId = decodedToken.id
+
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if (blogToDelete.user.toString() !== userId) {
+      console.log('ei mennyt läpi!!!')
+      return response.status(401).json({
+        error: 'deletion not authorized for this user'
+      })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     return response.status(204).end()
   } catch (exception) {
